refactor(ui): migrate app shell to TypeScript
- App.tsx: typed Theme/Navigation/Drawer/Subdrawer context providers via ThemeContextValue/NavigationContextValue/DrawerContextValue
- Main.tsx: typed checkAccess with Role + CrudPermission narrowing, Mongo roleId selector
- Login.tsx: typed LoginValues form shape
- Suspense.tsx: typed children prop
- ProfilePictureInput.tsx: typed FormInstance + UploadFile from antd
