+++ imports/ui/profile-picture-input/ProfilePictureInput.tsx
@@ -1,18 +1,19 @@
 import { App, Spin, Upload } from 'antd';
+import type { FormInstance } from 'antd';
+import type { UploadFile } from 'antd/es/upload/interface';
 import { Meteor } from 'meteor/meteor';
-import PropTypes from 'prop-types';
 import React, { useEffect, useState } from 'react';
 
-export async function turnImageFileToBase64(file) {
+export async function turnImageFileToBase64(file: Blob): Promise<string> {
   return new Promise((resolve, reject) => {
     const reader = new FileReader();
     reader.readAsDataURL(file);
-    reader.onload = () => resolve(reader.result);
+    reader.onload = () => resolve(reader.result as string);
     reader.onerror = error => reject(error);
   });
 }
 
-export async function turnBase64ToImage(base64) {
+export async function turnBase64ToImage(base64: string): Promise<HTMLImageElement> {
   return new Promise((resolve, reject) => {
     const img = new Image();
     img.onload = () => resolve(img);
@@ -21,19 +22,26 @@ export async function turnBase64ToImage(base64) {
   });
 }
 
-export default function ProfilePictureInput({ fileList, setFileList, form, profilePictureId }) {
+interface ProfilePictureInputProps {
+  fileList?: UploadFile[];
+  setFileList: (files: UploadFile[]) => void;
+  form: FormInstance;
+  profilePictureId?: string;
+}
+
+export default function ProfilePictureInput({ fileList, setFileList, form, profilePictureId }: ProfilePictureInputProps) {
   const { notification } = App.useApp();
   const [loading, setLoading] = useState(false);
-  const [imageSrc, setImageSrc] = useState(null);
+  const [imageSrc, setImageSrc] = useState<string | null>(null);
 
-  async function uploadImage(file) {
+  async function uploadImage(file: Blob) {
     setLoading(true);
     const base64 = await turnImageFileToBase64(file);
     setImageSrc(base64);
     Meteor.callAsync('profilePictures.insert', { value: base64 })
-      .catch(error => {
+      .catch((error: Meteor.Error) => {
         notification.error({
-          message: error.error,
+          message: error.error as string,
           description: error.message,
         });
       })
@@ -52,9 +60,9 @@ export default function ProfilePictureInput({ fileList, setFileList, form, profi
             setImageSrc(res[0].value);
           }
         })
-        .catch(error => {
+        .catch((error: Meteor.Error) => {
           notification.error({
-            message: error.error,
+            message: error.error as string,
             description: error.message,
           });
         })
@@ -69,8 +77,14 @@ export default function ProfilePictureInput({ fileList, setFileList, form, profi
       accept="image/*"
       directory={false}
       fileList={fileList}
-      beforeUpload={(_, fileList) => setFileList(fileList)}
-      customRequest={() => uploadImage(fileList[0])}
+      beforeUpload={(_, list) => {
+        setFileList(list);
+      }}
+      customRequest={() => {
+        if (fileList && fileList[0]) {
+          uploadImage(fileList[0] as unknown as Blob);
+        }
+      }}
     >
       <Spin spinning={loading}>
         {imageSrc ? (
@@ -82,9 +96,3 @@ export default function ProfilePictureInput({ fileList, setFileList, form, profi
     </Upload.Dragger>
   );
 }
-ProfilePictureInput.propTypes = {
-  fileList: PropTypes.array,
-  setFileList: PropTypes.func,
-  form: PropTypes.object,
-  profilePictureId: PropTypes.string,
-};
